pyproject.toml
[project]
name = "ecommerce-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "deprecated>=1.2.18",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "lxml>=6.0.1",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "pandas>=2.3.2",
    "playwright>=1.55.0",
    "prometheus-client>=0.22.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-json-logger>=3.3.0",
    "requests>=2.32.5",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]
